add screens and models, WIP adding app state.  State is implemented as an inheritedWidget within stateful widget.  Can access state using 'of' method.  Will use for cleaning auth implementation, which is now spread across multiple files
